Move phabricator into repo-updater
