+++ cmd/repo-updater/shared/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
     deps = [
         "//cmd/frontend/envvar",
         "//cmd/frontend/globals",
+        "//cmd/repo-updater/internal/phabricator",
         "//cmd/repo-updater/internal/repoupdater",
         "//cmd/repo-updater/internal/scheduler",
         "//internal/actor",
